prerendering + refactoring
