+++ src/routes/+page.js
@@ -1,0 +1,1 @@
+export const prerender = 'auto';
