+++ src/routes/api/background.svg/Background.svelte
